Initial code version
This is the initial configuration to explore observability capabilities
of Dynatrace Platform. Dynatrace collects data from a GKE cluster on
which there is a microservice based application installed.

There is step by step described all necessary configurations to be able
to bring up necessary components. Also, an utility script to
automatically deploy and configure necessary components